backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
deep-translator==1.11.4
langdetect==1.0.9
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.23
databases==0.8.0
aiosqlite==0.19.0